server/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
openai
httpx
pypdf
python-multipart
